archive done

+++ app/Http/Controllers/Admin/UserTypesController.php
@@ -115,6 +115,10 @@ class UserTypesController extends ConferenceBaseController
      */
     public function update(Requests\UserTypeRequest $request, UserType $type)
     {
+        if ($type->is_default) {
+            return redirect()->back()->with('error', 'is-default');
+        }
+
         DB::transaction(function () use ($request, $type) {
             $type->update($request->except('access'));
             $type->access()->delete();
